QSC-Node/requirements.txt
# Fixing sphinx version due to https://github.com/sphinx-doc/sphinx/issues/3976
# sudo apt -y install swig3.0 python3-dev build-essential cmake ninja-build pkg-config libhwloc-dev libboost-dev libssl-dev
# cd src/qsc/crypto/xmssmt/xmss-reference 
# make
# cd ..
# pip install Cython
# python setup.py install

#setuptools==50.3.2 # It needs to be installed manually in some cases
plyvel>=1.2.0,<=1.4.0
ntplib==0.3.4
Twisted==20.3.0
colorlog==3.1.0
simplejson==3.11.1
PyYAML==6.0 #5.3.1
grpcio-tools>=1.9.0,<=1.50.0
grpcio>=1.9.0,<=1.50.0
google-api-python-client==1.8.3
google-auth<2.0dev,>=1.21.1
httplib2>=0.15.0
service_identity==17.0.0
protobuf>3.19.0,<=3.20.3
pyopenssl==17.5.0
six==1.13.0
click==8.0
pyqrllib>=1.2.3,<1.3.0
pyqryptonight>=0.99.9
pyqrandomx>=0.3.0,<1.0.0
Flask>=2.0.0,<=2.2.2
json-rpc==1.13.0
idna==2.6
cryptography==2.3
mock>=2.0.0
daemonize==2.4.7

